# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Career Boost AI India API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithToken"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithToken"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List user resumes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /resumes/{id}/score:
    get:
      operationId: getResumeScore
      tags: [resumes]
      summary: Get ATS score for a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeScore"
  /resumes/analyze:
    post:
      operationId: analyzeResume
      tags: [resumes]
      summary: Analyze resume text and return score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeScore"
  /resumes/job-match:
    post:
      operationId: analyzeJobMatch
      tags: [resumes]
      summary: Analyze match between resume and job description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobMatchInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobMatchResult"

  # ─── Cover Letters ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── User schemas ─────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, isPremium, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        isPremium:
          type: boolean
        premiumExpiresAt:
          type: ["string", "null"]
        resumeCount:
          type: integer
        coverLetterCount:
          type: integer
        interviewCount:
          type: integer
        createdAt:
          type: string
    UserRegistration:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        mobile:
          type: string
        city:
          type: string
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UserWithToken:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...
    Resume:
      type: object
      required: [id, userId, title, template, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        fullName:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        projects:
          type: ["string", "null"]
        workExperience:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        languages:
          type: ["string", "null"]
        template:
          type: string
        generatedContent:
          type: ["string", "null"]
        atsScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ResumeInput:
      type: object
      required: [title, template]
      properties:
        title:
          type: string
        fullName:
          type: string
        mobile:
          type: string
        email:
          type: string
        city:
          type: string
        education:
          type: string
        college:
          type: string
        skills:
          type: string
        projects:
          type: string
        workExperience:
          type: string
        certifications:
          type: string
        languages:
          type: string
        template:
          type: string
    ResumeUpdate:
      type: object
      properties:
        title:
          type: string
        fullName:
          type: string
        mobile:
          type: string
        email:
          type: string
        city:
          type: string
        education:
          type: string
        college:
          type: string
        skills:
          type: string
        projects:
          type: string
        workExperience:
          type: string
        certifications:
          type: string
        languages:
          type: string
        template:
          type: string
        generatedContent:
          type: string
        atsScore:
          type: integer
    ResumeScore:
      type: object
      required: [overallScore, atsScore, skillsScore, formattingScore, suggestions]
      properties:
        overallScore:
          type: integer
        atsScore:
          type: integer
        skillsScore:
          type: integer
        formattingScore:
          type: integer
        suggestions:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
    ResumeAnalysisInput:
      type: object
      required: [resumeText]
      properties:
        resumeText:
          type: string
        jobRole:
          type: string
    JobMatchInput:
      type: object
      required: [resumeText, jobDescription]
      properties:
        resumeText:
          type: string
        jobDescription:
          type: string
    JobMatchResult:
      type: object
      required: [matchPercentage, missingSkills, suggestions]
      properties:
        matchPercentage:
          type: integer
        missingSkills:
          type: array
          items:
            type: string
        matchedSkills:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string

    # ─── Cover Letter schemas ─────────────────────────────────────────────────
...