# lib/api-zod/src/generated/types/jobMatchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Boost AI India API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobMatchResult {
  matchPercentage: number;
  missingSkills: string[];
  matchedSkills?: string[];
  suggestions: string[];
}