# lib/api-zod/src/generated/types/resumeScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Boost AI India API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeScore {
  overallScore: number;
  atsScore: number;
  skillsScore: number;
  formattingScore: number;
  suggestions: string[];
  strengths?: string[];
}