# lib/api-zod/src/generated/types/resumeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Boost AI India API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeUpdate {
  title?: string;
  fullName?: string;
  mobile?: string;
  email?: string;
  city?: string;
  education?: string;
  college?: string;
  skills?: string;
  projects?: string;
  workExperience?: string;
  certifications?: string;
  languages?: string;
  template?: string;
  generatedContent?: string;
  atsScore?: number;
}