# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Boost AI India API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  city?: string | null;
  isPremium: boolean;
  /** @nullable */
  premiumExpiresAt?: string | null;
  resumeCount?: number;
  coverLetterCount?: number;
  interviewCount?: number;
  createdAt: string;
}